scraper/requirements.txt
google-cloud-storage==2.16.0
google-cloud-secret-manager==2.20.0
requests==2.32.3
